Refactor SmokeTest so we can reuse methods
