+++ test/SmokeTest.php
@@ -4,57 +4,11 @@ namespace test\eLife\DummyApi;
 
 use PHPUnit_Framework_TestCase;
 use Symfony\Component\Finder\Finder;
-use Symfony\Component\HttpFoundation\Request;
 use Traversable;
 
 final class SmokeTest extends PHPUnit_Framework_TestCase
 {
     use SilexTestCase;
-
-    /**
-     * @test
-     */
-    public function it_can_be_pinged()
-    {
-        $response = $this->getApp()->handle(Request::create('/ping'));
-
-        $this->assertSame(200, $response->getStatusCode());
-        $this->assertSame('pong', $response->getContent());
-        $this->assertSame('text/plain; charset=UTF-8', $response->headers->get('Content-Type'));
-        $this->assertSame('must-revalidate, no-cache, no-store, private', $response->headers->get('Cache-Control'));
-    }
-
-    /**
-     * @test
-     * @dataProvider requestProvider
-     */
-    public function it_returns_valid_responses(Request $request, $contentType, int $statusCode = 200, $warning = [])
-    {
-        $response = $this->getApp()->handle($request);
-
-        if (in_array('--debug', $_SERVER['argv'], true) && 500 === $response->getStatusCode()) {
-            $json = json_decode($response->getContent(), true);
-            if (isset($json['exception'])) {
-                $this->fail($json['exception']);
-            }
-            $this->fail($json);
-        }
-
-        $this->assertSame($statusCode, $response->getStatusCode(), $response->getContent());
-        if (is_array($contentType)) {
-            $this->assertContains($response->headers->get('Content-Type'), $contentType);
-        } else {
-            $this->assertSame($contentType, $response->headers->get('Content-Type'));
-        }
-        if (strpos('+json', $response->headers->get('Content-Type'))) {
-            $this->assertTrue(is_array(json_decode($response->getContent(), true)), 'Does not contain a JSON response');
-        }
-        if (!empty($warning[$response->headers->get('Content-Type')])) {
-            $this->assertSame($warning[$response->headers->get('Content-Type')], $response->headers->get('Warning'));
-        } else {
-            $this->assertNull($response->headers->get('Warning'));
-        }
-    }
 
     public function requestProvider() : Traversable
     {
@@ -492,10 +446,5 @@ final class SmokeTest extends PHPUnit_Framework_TestCase
             'application/problem+json',
             400,
         ];
-    }
-
-    private function createRequest(string $uri, string $type = '*/*') : Request
-    {
-        return Request::create($uri, 'GET', [], [], [], ['HTTP_ACCEPT' => $type]);
     }
 }
